Refactor icons imports / add media controller
