+++ components/Layout/ExtensionSidebar.tsx
@@ -1,8 +1,7 @@
 import { styled } from '../../theme'
-import { ErrorIcon } from 'icons/Error'
+import { ErrorIcon, UpRightArrowIcon } from 'icons'
 import { __TEST_MODE__ } from '../../util/constants'
 import { Button, Text } from 'components'
-import { UpRightArrowIcon } from '../../icons/UpRightArrow'
 
 export const ExtensionSidebar = () => {
   return (
