+++ components/TokenSelectList.tsx
@@ -3,7 +3,7 @@ import { styled } from 'theme'
 import { TokenInfo } from 'hooks/useTokenList'
 import { Text } from 'components/Text'
 import { formatTokenBalance } from 'util/conversion'
-import { RejectIcon } from 'icons/Reject'
+import { RejectIcon } from 'icons'
 import { useTokenBalance } from 'hooks/useTokenBalance'
 import { useIBCTokenBalance } from 'hooks/useIBCTokenBalance'
 import { ButtonForWrapper } from './Button'
