+++ features/liquidity/components/AprPill.tsx
@@ -1,5 +1,5 @@
-import { ArrowUpIcon } from 'icons/ArrowUp'
-import { styled } from '../../../theme'
+import { ArrowUpIcon } from 'icons'
+import { styled } from 'theme'
 import { Inline, Text } from 'components'
 
 export const AprPill = ({ value, ...props }) => (
