+++ features/liquidity/components/UnbondingLiquidityStatusList.tsx
@@ -20,9 +20,7 @@ import {
 } from 'util/conversion'
 import { useClaimTokens } from '../hooks'
 import { toast } from 'react-hot-toast'
-import { Valid } from 'icons/Valid'
-import { Error } from 'icons/Error'
-import { UpRightArrow } from 'icons/UpRightArrow'
+import { Valid, Error, UpRightArrow } from 'icons'
 import { UnbondingLiquidityCard } from './UnbondingLiquidityCard'
 import { TokenInfo } from 'hooks/useTokenList'
 import { useRefetchQueries } from 'hooks/useRefetchQueries'
