+++ features/swap/components/TokenSelector.tsx
@@ -2,7 +2,7 @@ import { useTokenBalance } from 'hooks/useTokenBalance'
 import { styled } from 'theme'
 import React, { useRef, useState } from 'react'
 import { TokenOptionsList } from './TokenOptionsList'
-import { Union } from 'icons/Union'
+import { Union } from 'icons'
 import { useOnClickOutside } from 'hooks/useOnClickOutside'
 import { SelectorToggle } from './SelectorToggle'
 import { SelectorInput } from './SelectorInput'
