+++ icons/index.ts
@@ -1,3 +1,29 @@
-export { Address } from './Address'
-export { ArrowUp } from './ArrowUp'
-export { Open } from './Open'
+export * from './Address'
+export * from './ArrowUp'
+export * from './Chevron'
+export * from './Connect'
+export * from './Copy'
+export * from './Discord'
+export * from './DoubleArrow'
+export * from './Error'
+export * from './Exchange'
+export * from './Feedback'
+export * from './Github'
+export * from './Info'
+export * from './Logo'
+export * from './LogoBrighter'
+export * from './LogoText'
+export * from './Logout'
+export * from './Moon'
+export * from './Multisig'
+export * from './Open'
+export * from './Plus'
+export * from './Reject'
+export * from './Search'
+export * from './Shares'
+export * from './Telegram'
+export * from './Twitter'
+export * from './Union'
+export * from './UpRightArrow'
+export * from './Valid'
+export * from './Wallet'
